+++ design-system/packages/checkbox/src/index.ts
@@ -1,3 +1,4 @@
 export { Checkbox } from './Checkbox';
+export { CheckboxGroup } from './CheckboxGroup';
 
-export type { CheckboxProps } from './types';
+export type { CheckboxProps, CheckboxGroupProps } from './types';