+++ design-system/packages/radio/src/index.ts
@@ -1,0 +1,4 @@
+export { Radio } from './Radio';
+export { RadioGroup } from './RadioGroup';
+
+export type { RadioProps, RadioGroupProps } from './types';