+++ apps/astro/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @keystatic/test-astro
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [d196cb0]
+  - @keystatic/core@0.0.45
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [9a5d260]
+  - @keystatic/core@0.0.44
+
 ## 0.0.7
 
 ### Patch Changes