+++ packages/keystatic/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @keystatic/core
 
+## 0.0.45
+
+### Patch Changes
+
+- d196cb0: Change suggested config file name
+
+## 0.0.44
+
+### Patch Changes
+
+- 9a5d260: Fix showing "repo not found error" when changing between different
+  Keystatic apps that use different repos.
+
 ## 0.0.43
 
 ### Patch Changes